requirements.txt
configparser
pymongo==3.10.1
dnspython==1.16.0
python-telegram-bot>=10.0.2
requests>=2.20.0
beautifulsoup4==4.6.0
dataclasses==0.7
lxml==4.2.4
apscheduler
